Fix pnpm check: split proposals.ts below 1000-line limit
Move getProposalCoSpeakers and getPresentationUploader to
proposals-speaker-profile.ts, and markProposalStatus / softDeleteProposal
to a new proposals-admin.ts. Re-export from proposals.ts for backwards
compat so no call sites need updating.
